tmf: fix test failures due execution timeouts
The timeout deadline used to be calculated using integer instead of
long so that an overflow occurred and the actual deadline was much
shorter than intended.

Reviewed-on: https://git.eclipse.org/r/20224
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Hudson CI

+++ org.eclipse.linuxtools.tmf.ui.tests/shared/org/eclipse/linuxtools/tmf/ui/tests/shared/ProjectModelTestData.java
@@ -46,7 +46,7 @@ public class ProjectModelTestData {
     /* Maximum number of thread delays the main thread will do before timing out */
     private static final int DELAY_COUNTER = 10;
     /* Default delay time when having the main thread sleep. */
-    private static final int DEFAULT_DELAY = 500;
+    private static final long DEFAULT_DELAY = 500;
 
     /** Default test project name */
     public static final String PROJECT_NAME = "Test_Project";
@@ -210,7 +210,7 @@ public class ProjectModelTestData {
     public static void delayUntilTraceOpened(final ITmfProjectModelElement projectElement) throws TimeoutException {
         if (projectElement instanceof TmfTraceElement) {
             TmfTraceElement traceElement = (TmfTraceElement) projectElement;
-            final long deadline = System.nanoTime() + (DELAY_COUNTER * DEFAULT_DELAY * 1000000);
+            final long deadline = System.nanoTime() + (DELAY_COUNTER * DEFAULT_DELAY * 1000000L);
             do {
                 delayThread(DEFAULT_DELAY);
                 if (traceElement.getTrace() != null) {
